backend/requirements.txt
fastapi==0.132.0
uvicorn[standard]>=0.29.0,<1.0.0
SQLAlchemy>=2.0.0,<3.0.0
psycopg2-binary>=2.9.0,<3.0.0
pydantic>=2.0.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
email-validator>=2.0.0,<3.0.0
